% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{alternative_validation}
\alias{alternative_validation}
\title{Alternative validation for large matrices}
\usage{
alternative_validation(abc_results, co_matrix, alpha, correction)
}
\description{
Alternative validation for large matrices
}
\keyword{internal}
