% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_bibliometrics}
\alias{calc_bibliometrics}
\title{Calculate basic bibliometric statistics}
\usage{
calc_bibliometrics(article_data, by_year = TRUE)
}
\arguments{
\item{article_data}{A data frame containing article data.}

\item{by_year}{Logical. If TRUE, calculates statistics by year.}
}
\value{
A list containing bibliometric statistics.
}
\description{
This function calculates basic bibliometric statistics from article data.
}
\examples{
\dontrun{
stats <- calc_bibliometrics(article_data)
}
}
