% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{diversify_b_terms}
\alias{diversify_b_terms}
\title{Enforce diversity by selecting top connections from each B term}
\usage{
diversify_b_terms(results, max_per_group = 3)
}
\arguments{
\item{results}{Data frame with ABC model results}

\item{max_per_group}{Maximum number of results to keep per B term}
}
\value{
Data frame with diverse results
}
\description{
Enforce diversity by selecting top connections from each B term
}
\keyword{internal}
