% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{has_general_biomedical_characteristics}
\alias{has_general_biomedical_characteristics}
\title{Check for general biomedical characteristics}
\usage{
has_general_biomedical_characteristics(term)
}
\description{
Check for general biomedical characteristics
}
\keyword{internal}
