% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_df}
\alias{list_to_df}
\title{Convert a list of articles to a data frame}
\usage{
list_to_df(articles)
}
\arguments{
\item{articles}{A list of articles, each containing metadata.}
}
\value{
A data frame containing article metadata.
}
\description{
This function converts a list of articles to a data frame.
}
\keyword{internal}
