% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_results}
\alias{merge_results}
\title{Merge multiple search results}
\usage{
merge_results(..., remove_duplicates = TRUE)
}
\arguments{
\item{...}{Data frames containing search results.}

\item{remove_duplicates}{Logical. If TRUE, removes duplicate articles.}
}
\value{
A merged data frame.
}
\description{
This function merges multiple search results into a single data frame.
}
\examples{
\dontrun{
merged_results <- merge_results(results1, results2, results3)
}
}
