% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{perm_test_abc}
\alias{perm_test_abc}
\title{Perform randomization test for ABC model}
\usage{
perm_test_abc(abc_results, co_matrix, n_permutations = 1000, alpha = 0.05)
}
\arguments{
\item{abc_results}{A data frame containing ABC results.}

\item{co_matrix}{The co-occurrence matrix used to generate the ABC results.}

\item{n_permutations}{Number of permutations to perform.}

\item{alpha}{Significance level.}
}
\value{
A data frame with ABC results and permutation-based significance measures.
}
\description{
This function assesses the significance of ABC model results through randomization.
It generates a null distribution by permuting the co-occurrence matrix.
}
\examples{
\dontrun{
randomized_results <- perm_test_abc(abc_results, co_matrix, n_permutations = 1000)
}
}
