% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{query_external_api}
\alias{query_external_api}
\title{Query external biomedical APIs to validate entity types}
\usage{
query_external_api(term, claimed_type)
}
\arguments{
\item{term}{Character string, the term to validate}

\item{claimed_type}{Character string, the claimed entity type}
}
\value{
Logical indicating if the term was found in the appropriate database
}
\description{
Query external biomedical APIs to validate entity types
}
