% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{remove_ac_terms}
\alias{remove_ac_terms}
\title{Remove A and C terms that appear as B terms}
\usage{
remove_ac_terms(results)
}
\arguments{
\item{results}{Data frame with ABC model results}
}
\value{
Data frame with A and C terms removed from B terms
}
\description{
Remove A and C terms that appear as B terms
}
\keyword{internal}
