% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{shadowtext}
\alias{shadowtext}
\title{Helper function to draw text with a shadow/background}
\usage{
shadowtext(
  x,
  y,
  labels,
  col = "black",
  bg = "white",
  pos = NULL,
  offset = 0.5,
  cex = 1,
  ...
)

shadowtext(
  x,
  y,
  labels,
  col = "black",
  bg = "white",
  pos = NULL,
  offset = 0.5,
  cex = 1,
  ...
)
}
\description{
Helper function to draw text with a shadow/background

Helper function to draw text with a shadow/background
}
\keyword{internal}
