% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{standard_validation}
\alias{standard_validation}
\title{Standard validation method using hypergeometric tests}
\usage{
standard_validation(abc_results, co_matrix, alpha, correction)
}
\description{
Standard validation method using hypergeometric tests
}
\keyword{internal}
