% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{validate_biomedical_entity}
\alias{validate_biomedical_entity}
\title{Validate biomedical entities using BioBERT or other ML models}
\usage{
validate_biomedical_entity(term, claimed_type)
}
\arguments{
\item{term}{Character string, the term to validate}

\item{claimed_type}{Character string, the claimed entity type}
}
\value{
Logical indicating if the term is validated
}
\description{
Validate biomedical entities using BioBERT or other ML models
}
