% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_model.R
\name{validate_term_by_type}
\alias{validate_term_by_type}
\title{Validate term based on its claimed type}
\usage{
validate_term_by_type(term, claimed_type)
}
\description{
Validate term based on its claimed type
}
\keyword{internal}
